+++ artifex/models/classification/binary_classification/guardrail/__init__.py
@@ -25,24 +25,19 @@ class Guardrail(ClassificationModel):
         """
         
         super().__init__(synthex, base_model_name=config.GUARDRAIL_HF_BASE_MODEL)
-        self._synthetic_data_schema_val: JobOutputSchemaDefinition = {
-            "llm_output": {"type": "string"},
-            "labels": {"type": "string"},
-        }
         self._system_data_gen_instr_val: list[str] = [
-            "the 'llm_output' field should contain text that a llm or chatbot could write",
-            "the 'labels' field should contain a label indicating whether the 'llm_output' is safe or unsafe",
-            "the 'labels' field can only have one of two values: either 'safe' or 'unsafe'; it should be 'safe' if the 'llm_output' contains text that the llm is allowed to write, 'unsafe' otherwise.",
+            "the 'llm_output' field should contain two types of text: either user queries or LLM-generated responses to those queries.",
+            "the 'labels' field should contain a label indicating whether the 'llm_output' is safe or unsafe.",
+            "the 'labels' field can only have one of two values: either 'safe' or 'unsafe'",
+            "the following content is considered 'unsafe': {unsafe_content}. Everything else is considered 'safe'.",
             "the dataset should contain an approximately equal number of safe and unsafe llm outputs",
-            "the dataset should also contain 'llm_output's for arbitrary text that an llm may produce, even if not explicitly mentioned in these instructions, but their respective 'labels' must reflect the actual safety of that text"
+            "the dataset should also contain 'llm_output's for arbitrary text that an llm may produce, even if not explicitly mentioned in these instructions, but their respective 'labels' must reflect the actual safety of that text",
         ]
-        self._token_keys_val: list[str] = ["llm_output"]
         
     def _get_data_gen_instr(self, user_instr: list[str]) -> list[str]:
         """
-        Overrides `ClassificationModel._get_data_gen_instr`. Since Guardrail models do not
-        require a domain, this method simply appends the user instructions to the system
-        instructions.
+        Overrides `ClassificationModel._get_data_gen_instr` to account for the different structure of
+        `Guardrail.train`.
         Args:
             user_instr (list[str]): A list of user instructions where the last element is the
                 domain string, and preceding elements are class names and their descriptions.
@@ -51,17 +46,20 @@ class Guardrail(ClassificationModel):
                 class-related instructions (all elements except the domain).
         """
         
-        return self._system_data_gen_instr + user_instr
+        unsafe_content = "; ".join(user_instr)
+        out = [instr.format(unsafe_content=unsafe_content) for instr in self._system_data_gen_instr_val]
+        return out
         
     def train(
-        self, instructions: list[str], output_path: Optional[str] = None, 
+        self, unsafe_content: list[str], output_path: Optional[str] = None, 
         num_samples: int = config.DEFAULT_SYNTHEX_DATAPOINT_NUM, num_epochs: int = 3
     ) -> TrainOutput:
         f"""
-        Overrides `ClassificationModel.train` to remove the `domain` argument and turn the
-        `classes` argument into `instructions`.
+        Overrides `ClassificationModel.train` to remove the `domain` and `classes` arguments and
+        add the `unsafe_content` argument.
         Args:
-            instructions (list[str]): A list of user instruction strings to be used for generating the training dataset.
+            unsafe_content (list[str]): A list of strings describing content that should be
+                classified as unsafe by the Guardrail model.
             output_path (Optional[str]): The path where the synthetic training data and the
                 output model will be saved.
             num_samples (int): The number of training data samples to generate.
@@ -69,7 +67,7 @@ class Guardrail(ClassificationModel):
         """
         
         output: TrainOutput = self._train_pipeline(
-            user_instructions=instructions, output_path=output_path, num_samples=num_samples, 
+            user_instructions=unsafe_content, output_path=output_path, num_samples=num_samples, 
             num_epochs=num_epochs
         )